django_spravka11/requirements.txt
Django==1.7.11
Pillow==2.7.0
django-filebrowser==3.5.7
django-grappelli==2.6.5
django-haystack==2.3.1
django-mptt==0.6.1
pysolr==3.3.0
requests==2.6.0
sorl-thumbnail==12.2
virtualenv==1.11.6
django-import-export==0.2.6
PyMySQL==0.6.6
mysqlclient
django-disqus